refactor: isolate matchmaking network logic
